requirements.txt
flask>=2.0.0
gitpython>=3.1.0
python-dotenv>=0.19.0
requests>=2.26.0
psutil>=5.9.0  # For disk space checking 